create parameter module and remove condition file
